Remove CARS/CRAB
I just don't think this counts. This tool exists solely to allow counties to report data to the state. Absent the need for state reporting, there would be no need for this to exist. I don't think this is different than any government form for e.g. grantees to complete.