Create basis
Create a new collaborative file for Bay Area Spatial Information Systems (from Airtable) **note - I put in the command to add a picture, but I do not have access to add the picture from Airtable. I have the screenshot on file, so I can send it to someone to add, but I currently am unable to upload it to GitHub.